+++ Mindkeepr/serializers/Serializersold.py
@@ -1,0 +1,25 @@
+
+
+#class BorrowingSerializer(serializers.HyperlinkedModelSerializer):
+#    id = serializers.IntegerField()
+#
+#    location_source = LocationSerializer()
+#    element = ElementMinSerializer()
+#
+#    class Meta:
+#        model = BorrowEvent
+#        depth = 1
+#        fields = ("id", "quantity", "recording_date", "scheduled_return_date",
+#                  "location_source",
+#                  "element",
+#                  "comment")
+
+
+
+
+
+
+
+
+
+
