+++ Mindkeepr/serializers/attribute.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+from Mindkeepr.models import Attribute
+
+
+class AttributeSerializer(serializers.HyperlinkedModelSerializer):
+
+    class Meta:
+        model = Attribute
+        fields = ("id", "name", "value")
+        depth = 2
