+++ Mindkeepr/serializers/events/__init__.py
@@ -1,0 +1,11 @@
+#from .borrow_event import *
+#from .buy_event import *
+#from .consume_event import *
+#from .event import *
+#from .incident_event import *
+#from .maintenance_event import *
+#from .move_event import *
+#from .return_event import *
+#from .sell_event import *
+#from .unuse_event import *
+#from .use_event import *
