+++ Mindkeepr/serializers/fields.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+
+class RecursiveField(serializers.Serializer):
+    def to_representation(self, value):
+        serializer = self.parent.parent.__class__(value, context=self.context)
+        return serializer.data
