+++ Mindkeepr/views/location_view.py
@@ -3,7 +3,7 @@ from django.views.generic.edit import CreateView, UpdateView, DeleteView
 from django.contrib.auth.mixins import PermissionRequiredMixin
 from Mindkeepr.forms import LocationForm
 from Mindkeepr.models import Location
-from Mindkeepr.Serializers import LocationSerializer, LocationFullSerializer
+from Mindkeepr.serializers.location import LocationSerializer, LocationFullSerializer
 from django.urls import reverse_lazy
 from django.views.generic.list import ListView
 from rest_framework import viewsets
