+++ Mindkeepr/views/project_view.py
@@ -7,7 +7,7 @@ from django.contrib.auth.mixins import PermissionRequiredMixin
 from rest_framework import viewsets
 from Mindkeepr.forms import ProjectForm
 from Mindkeepr.models import Project
-from Mindkeepr.Serializers import ProjectSerializer
+from Mindkeepr.serializers.project import ProjectSerializer
 from django.views.generic.list import ListView
 from .search import searchFilter
